feat: View에 대한 인터페이스 적용
