+++ src/main/kotlin/Application.kt
@@ -1,5 +1,11 @@
 import controller.OmokController
+import domain.rule.RenjuRule
+import view.OmokInputView
+import view.OmokOutputView
 
 fun main() {
-    OmokController().start()
+    OmokController(
+        OmokInputView(),
+        OmokOutputView()
+    ).start(RenjuRule())
 }
