+++ src/main/kotlin/controller/OmokController.kt
@@ -1,12 +1,15 @@
 package controller
 
 import domain.game.Omok
-import domain.rule.RenjuRule
+import domain.rule.OmokRule
 import view.InputView
 import view.OutputView
 
-class OmokController {
-    fun start() {
-        Omok(OutputView(), InputView(), RenjuRule()).run()
+class OmokController(
+    private val inputView: InputView,
+    private val outputView: OutputView,
+) {
+    fun start(omokRule: OmokRule) {
+        Omok(inputView, outputView, omokRule).run()
     }
 }
